src/requirements.txt
gunicorn==20.0.4
six==1.15.0
beautifulsoup4==4.8.0
dataclasses==0.6
filelock==3.0.12
firebase-admin==4.5.1
firestore==0.0.8
Flask==1.1.2
gcloud==0.17.0
google==3.0.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.26.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-cloud-core==1.6.0
google-cloud-error-reporting==1.1.0
google-cloud-firestore==2.0.2
google-cloud-logger==0.2.1
google-cloud-logging==2.2.0
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
lxml==4.5.2
pyrebase5==5.0.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
Wand==0.6.3
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.3.6
